feat: add validation in categorias personalizadas service

+++ services/categoriasPersonalizadas.api.service.js
@@ -35,6 +35,10 @@ const getCategoriaPersonalizadaById = async (id) => {
 const deleteCategoriaPersonalizada = async (id) => {
     try {
         const connection = await connectToDatabase();
+        const [categoria] = await connection.execute('SELECT * FROM categoriasPersonalizadas WHERE idCategoriaPersonalizada = ?', [id]);
+        if (categoria.length === 0) {
+            throw new Error('Categoría personalizada no encontrada');
+        }
         await connection.execute('DELETE FROM categoriasPersonalizadas WHERE idCategoriaPersonalizada = ?', [id]);
         await connection.end();
     } catch (err) {
@@ -45,6 +49,10 @@ const deleteCategoriaPersonalizada = async (id) => {
 const updateCategoriaPersonalizada = async (id, nombre) => {
     try {
         const connection = await connectToDatabase();
+        const [categoria] = await connection.execute('SELECT * FROM categoriasPersonalizadas WHERE idCategoriaPersonalizada = ?', [id]);
+        if (categoria.length === 0) {
+            throw new Error('Categoría personalizada no encontrada');
+        }
         await connection.execute('UPDATE categoriasPersonalizadas SET nombre = ? WHERE idCategoriaPersonalizada = ?', [nombre, id]);
         await connection.end();
     } catch (err) {
